requirements.txt
flask_socketio
flask
pywebview
kthread
psutil
